Rename ListenerHelper to ListenerBus and the original ListenerBus to AsynchronousListenerBus
